Fix: changed the numeric type of multiple files necessary (float  ->  int)

+++ print_length_of_odd_numbers_from_ten_user_inputs.py
@@ -7,7 +7,7 @@ odd_numbers = []
 # While loop for 10 inputs
 while counter < 10:
     counter += 1
-    ten_numbers = float(input("Enter 10 numbers: "))
+    ten_numbers = int(input("Enter 10 numbers: "))
 
     # Determines odd numbers
     if ten_numbers % 2 != 0:
